Update rerun integration for current SDK

+++ include/aria_viz/visualizer_rerun.h
@@ -82,7 +82,7 @@ class VisualizerRerun : public Visualizer {
     // 3. Layout and render to memory (PNG format)
     gvLayout(gvc, g, "dot");
     char* data = nullptr;
-    size_t length = 0;
+    unsigned int length = 0;
     gvRenderData(gvc, g, "png", &data, &length);
 
     // 4. Convert rendered data to OpenCV Mat
